# docs/swagger.yaml
basePath: /
definitions:
  api.Accept:
    properties:
      flag:
        type: boolean
    type: object
  api.Audit:
    properties:
      limit:
        $ref: '#/definitions/api.Limit'
    type: object
  api.Destination:
    properties:
      address:
        type: string
      invert:
        type: string
    type: object
  api.Drop:
    properties:
      flag:
        type: boolean
    type: object
  api.ForwardPort:
    properties:
      port:
        type: string
      protocol:
        type: string
      toaddr:
        type: string
      toport:
        type: string
    type: object
  api.IcmpBlock:
    properties:
      name:
        type: string
    type: object
  api.IcmpType:
    properties:
      name:
        type: string
    type: object
...
  api.Limit:
    properties:
      value:
        type: string
    type: object
  api.Log:
    properties:
      level:
        type: string
      limit:
        $ref: '#/definitions/api.Limit'
      prefix:
        type: string
    type: object
  api.Mark:
    properties:
      limit:
        $ref: '#/definitions/api.Limit'
      set:
        type: string
    type: object
  api.Port:
    properties:
      port:
        type: string
      protocol:
        type: string
    type: object
  api.Protocol:
    properties:
      value:
        type: string
    type: object
...
  api.Reject:
    properties:
      flag:
        type: boolean
    type: object
  api.Rule:
    properties:
      accept:
        $ref: '#/definitions/api.Accept'
      audit:
        $ref: '#/definitions/api.Audit'
      destination:
        $ref: '#/definitions/api.Destination'
      drop:
        $ref: '#/definitions/api.Drop'
      family:
        type: string
      forwardport:
        $ref: '#/definitions/api.ForwardPort'
      icmpblock:
        $ref: '#/definitions/api.IcmpBlock'
      icmptype:
        $ref: '#/definitions/api.IcmpType'
      limit:
        $ref: '#/definitions/api.Limit'
      log:
        $ref: '#/definitions/api.Log'
      mark:
        $ref: '#/definitions/api.Mark'
      port:
        $ref: '#/definitions/api.Port'
      protocol:
        $ref: '#/definitions/api.Protocol'
      reject:
        $ref: '#/definitions/api.Reject'
      service:
        items:
          type: string
        type: array
      source:
        $ref: '#/definitions/api.Source'
      value:
        $ref: '#/definitions/api.Value'
    type: object
...
  api.Source:
    properties:
      address:
        type: string
      invert:
        type: string
      ipset:
        type: string
      mac:
        type: string
    type: object
...
  api.Value:
    properties:
      value:
        type: string
    type: object
...
  query.BatchForwardQuery:
    properties:
      delay:
        type: integer
      forwards:
        items:
          $ref: '#/definitions/query.ForwardQuery'
        type: array
    type: object
  query.BatchPortQuery:
    properties:
      delay:
        type: integer
      ports:
        items:
          $ref: '#/definitions/query.PortQuery'
        type: array
    type: object
  query.BatchRichQuery:
    properties:
      delay:
        type: integer
      richs:
        items:
          $ref: '#/definitions/query.RichQuery'
        type: array
    type: object
...
  query.ForwardQuery:
    properties:
      forward:
        $ref: '#/definitions/api.ForwardPort'
      ip:
        type: string
      timeout:
        type: integer
      zone:
        type: string
    required:
    - forward
    - ip
    type: object
...
  query.IDQuery:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  query.ListQuery:
    properties:
      id:
        type: integer
      limit:
        type: integer
      offset:
        type: integer
      simple:
        type: integer
      sort:
        type: string
      title:
        type: string
    type: object
  query.PortEditQuery:
    properties:
      id:
        type: integer
      port:
        type: integer
      protocol:
        type: string
      template_id:
        type: integer
    required:
    - port
    - protocol
    - template_id
    type: object
  query.PortQuery:
    properties:
      ip:
        type: string
      port:
        $ref: '#/definitions/api.Port'
      timeout:
        type: integer
      zone:
        type: string
    required:
    - ip
    - port
    type: object
  query.Query:
    properties:
      forward:
        $ref: '#/definitions/api.ForwardPort'
      ip:
        type: string
      port:
        $ref: '#/definitions/api.Port'
      rich:
        $ref: '#/definitions/api.Rule'
      service:
        type: string
      timeout:
        type: integer
      zone:
        type: string
    required:
    - ip
    type: object
...
  query.RichEditQuery:
    properties:
      action:
        type: string
      destination:
        type: string
      family:
        type: string
      id:
        type: integer
      limit:
        type: integer
      limit_unit:
        type: string
      port:
        type: string
      protocol:
        type: string
      source:
        type: string
      template_id:
        type: integer
    required:
    - action
    - template_id
    type: object
  query.RichQuery:
    properties:
      ip:
        type: string
      rich:
        $ref: '#/definitions/api.Rule'
      timeout:
        type: integer
      zone:
        type: string
    required:
    - ip
    - rich
    type: object
...
host: localhost:2952
info:
  contact: {}
  description: Uranus, distrubed firewall gateway.
  title: Uranus API
  version: 0.0.9
paths:
...
  /fw/template/port:
    delete:
      consumes:
      - application/json
      description: Delete port rule with rule id.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.IDQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Delete port rule with rule id.
      tags:
      - Template
    get:
      consumes:
      - application/json
      description: List port rules.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.ListQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: List port rules.
      tags:
      - Template
    post:
      consumes:
      - application/json
      description: Update port rule information with port rule id.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.PortEditQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update port rule information with port rule id.
      tags:
      - Template
    put:
      consumes:
      - application/json
      description: Create a new port rule to template.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.PortEditQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create a new port rule to template.
      tags:
      - Template
  /fw/template/rich:
    delete:
      consumes:
      - application/json
      description: Delete rich rule with rule id.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.IDQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Delete rich rule with rule id.
      tags:
      - Template
    get:
      consumes:
      - application/json
      description: List rich rules.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.ListQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: List rich rules.
      tags:
      - Template
    post:
      consumes:
      - application/json
      description: Update rich rule information with rich rule id.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.RichEditQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Update rich rule information with rich rule id.
      tags:
      - Template
    put:
      consumes:
      - application/json
      description: Create a new rich rule to template.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.RichEditQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create a new rich rule to template.
      tags:
      - Template
...
  /fw/v1/nat:
    delete:
      consumes:
      - application/json
      description: Remove a nat rule at firewalld runtime.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.Query'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Remove a nat rule at firewalld runtime.
      tags:
      - firewalld NAT
    get:
      consumes:
      - application/json
      description: Get nat rules at firewalld runtime
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.Query'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get nat rules at firewalld runtime.
      tags:
      - firewalld NAT
    put:
      consumes:
      - application/json
      description: Add a nat rule at firewall runtime.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.ForwardQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Add a nat rule at firewall runtime.
      tags:
      - firewalld NAT
  /fw/v1/ports:
    delete:
      consumes:
      - application/json
      description: Delete a firewalld port rule in runtime.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.PortQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a firewalld port rule in runtime.
      tags:
      - firewalld port
    get:
      consumes:
      - application/json
      description: Get port rule at firewalld runtime.
      parameters:
      - description: body
        in: query
        name: ip
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Get port rule at firewalld runtime.
      tags:
      - firewalld port
    put:
      consumes:
      - application/json
      description: Add port rule at firewalld runtime.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.Query'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Add port rule at firewalld runtime.
      tags:
      - firewalld port
  /fw/v1/rich:
    delete:
      consumes:
      - application/json
      description: Remove rich rule at firewalld runtimes.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.RichQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Remove rich rule at firewalld runtimes.
      tags:
      - firewalld rich
    get:
      consumes:
      - application/json
      description: Get rich rule list at firewalld runtimes.
      parameters:
      - description: body
        in: query
        name: ip
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Get rich rule list at firewalld runtimes.
      tags:
      - firewalld rich
    put:
      consumes:
      - application/json
      description: Add rich rule at firewalld runtimes.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.RichQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Add rich rule at firewalld runtimes.
      tags:
      - firewalld rich
...
  /fw/v2/nat:
    delete:
      consumes:
      - application/json
      description: Remove a nat rule at firewalld permanent.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.Query'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Remove a nat rule at firewalld permanent.
      tags:
      - firewalld NAT
    get:
      consumes:
      - application/json
      description: Get nat rules at firewalld permanent.
      parameters:
      - description: ip
        in: query
        name: ip
        required: true
        type: string
      - description: zone
        in: query
        name: zone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Get nat rules at firewalld permanent.
      tags:
      - firewalld NAT
    put:
      consumes:
      - application/json
      description: Add a nat rule at firewall permanent.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.ForwardQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Add a nat rule at firewall permanent.
      tags:
      - firewalld NAT
  /fw/v2/ports:
    delete:
      consumes:
      - application/json
      description: Delete a firewalld port rule on permanent.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.PortQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a firewalld port rule on permanent.
      tags:
      - firewalld port
    get:
      consumes:
      - application/json
      description: Get port rule on firewalld permanent.
      parameters:
      - description: body
        in: query
        name: ip
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Get port rule on firewalld permanent.
      tags:
      - firewalld port
    put:
      consumes:
      - application/json
      description: Add port rule on firewalld permanent.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.Query'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Add port rule on firewalld permanent.
      tags:
      - firewalld port
  /fw/v2/rich:
    delete:
      consumes:
      - application/json
      description: Remove rich rule on firewalld permanent.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.RichQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Remove rich rule on firewalld permanent.
      tags:
      - firewalld rich
    get:
      consumes:
      - application/json
      description: Get rich rule list on firewalld permanent.
      parameters:
      - description: body
        in: query
        name: ip
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Get rich rule list on firewalld permanent.
      tags:
      - firewalld rich
    put:
      consumes:
      - application/json
      description: Add rich on firewalld permanent.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.RichQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Add rich rule on firewalld permanent.
      tags:
      - firewalld rich
...
  /fw/v3/nat:
    put:
      consumes:
      - application/json
      description: Add a NAT rule on firewalld runtime with delay timer.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.BatchForwardQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Add a NAT rule on firewalld runtime with delay timer.
      tags:
      - firewalld NAT
  /fw/v3/ports:
    put:
      consumes:
      - application/json
      description: Add a port rule on firewalld runtime with delay timer.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.BatchPortQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Add a port rule on firewalld runtime with delay timer.
      tags:
      - firewalld port
  /fw/v3/ports/permanent:
    put:
      consumes:
      - application/json
      description: Add a port rule on firewalld permanent with delay timer.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.BatchPortQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Add a port rule on firewalld permanent with delay timer.
      tags:
      - firewalld port
  /fw/v3/rich:
    put:
      consumes:
      - application/json
      description: Add a rich rules on firewalld runtime with delay timer.
      parameters:
      - description: body
        in: body
        name: query
        schema:
          $ref: '#/definitions/query.BatchRichQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Add a rich rules on firewalld runtime with delay timer.
      tags:
      - firewalld rich
...
schemes:
- http
securityDefinitions:
  Bearer:
    description: 'Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"